feat: replace scoreboard with leaderboard, transfer agent tracking to reports
- rewrite scoreboard page as closed-cases leaderboard (3 teams, 5 weeks)
- extract agent tracking table + daily call entry into ScoreboardTracker component
- add ScoreboardTracker (with team stats) to the bottom of the reports page
